Force GriddedPSF to use the given psf data
CRDS epsf is oversampled. When creating GriddedPSF with the
accurate oversampling, GriddedPSF returns psfs that are at
the native pixel scale. Downstream use of the psfs then require
the psfs to be upscaled back to the oversampled size.

This is counter-intuitive since the original psfs were already at
the oversampled size. To ensure using the psfs as provided, the
option to lie to the GriddedPSF an oversampling is provided, enabling
GriddedPSF to use/return psfs at their originally provided scale.